feature complete

+++ build/backend/queryCmc.js
@@ -6,25 +6,47 @@ var fetch = require('node-fetch');
 var fs = require('fs');
 
 var queryCmc = function () {
-  var _ref = _asyncToGenerator( /*#__PURE__*/regeneratorRuntime.mark(function _callee2(db, url) {
+  var _ref = _asyncToGenerator( /*#__PURE__*/regeneratorRuntime.mark(function _callee2(db, url, debug) {
+    var response, coins;
     return regeneratorRuntime.wrap(function _callee2$(_context2) {
       while (1) {
         switch (_context2.prev = _context2.next) {
           case 0:
+            if (!debug) {
+              _context2.next = 4;
+              break;
+            }
+
             fs.readFile('./sample.json', function () {
               var _ref2 = _asyncToGenerator( /*#__PURE__*/regeneratorRuntime.mark(function _callee(err, data) {
+                var coins;
                 return regeneratorRuntime.wrap(function _callee$(_context) {
                   while (1) {
                     switch (_context.prev = _context.next) {
                       case 0:
-                        _context.next = 2;
-                        return db.updateCoins(JSON.parse(data));
+                        if (!err) {
+                          _context.next = 2;
+                          break;
+                        }
+
+                        throw new Error('Error reading sample data');
 
                       case 2:
-                        _context.next = 4;
+                        coins = JSON.parse(data);
+
+
+                        coins = coins.map(function (c) {
+                          c.price_usd *= Math.random();
+                          return c;
+                        });
+                        _context.next = 6;
+                        return db.updateCoins(coins);
+
+                      case 6:
+                        _context.next = 8;
                         return db.deleteOldPrices();
 
-                      case 4:
+                      case 8:
                       case 'end':
                         return _context.stop();
                     }
@@ -32,18 +54,34 @@ var queryCmc = function () {
                 }, _callee, undefined);
               }));
 
-              return function (_x3, _x4) {
+              return function (_x4, _x5) {
                 return _ref2.apply(this, arguments);
               };
             }());
+            _context2.next = 15;
+            break;
 
-            //const response = await fetch(url)
-            //const coins = await response.json()
-            //console.log('Fetched data for ' + coins.length.toString() + ' coins')
-            //await db.updateCoins(coins)
-            //await db.deleteOldPrices()
+          case 4:
+            _context2.next = 6;
+            return fetch(url);
 
-          case 1:
+          case 6:
+            response = _context2.sent;
+            _context2.next = 9;
+            return response.json();
+
+          case 9:
+            coins = _context2.sent;
+
+            console.log('Fetched data for ' + coins.length.toString() + ' coins');
+            _context2.next = 13;
+            return db.updateCoins(coins);
+
+          case 13:
+            _context2.next = 15;
+            return db.deleteOldPrices();
+
+          case 15:
           case 'end':
             return _context2.stop();
         }
@@ -51,7 +89,7 @@ var queryCmc = function () {
     }, _callee2, undefined);
   }));
 
-  return function queryCmc(_x, _x2) {
+  return function queryCmc(_x, _x2, _x3) {
     return _ref.apply(this, arguments);
   };
 }();
